pyproject.toml
[tool.poetry]
name = "aigc-game"
version = "0.1.0"
description = "An innovative game project leveraging AI agents for dynamic narrative and player interactions"
authors = ["Your Team"]
readme = "README.md"
packages = [{include = "aigc_game"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
# 原有依赖
openai = "^1.95.1"  # For LLM interactions
pydantic = "^2.10.6"  # For data validation
fastapi = "^0.109.0"  # For API interfaces
sqlalchemy = "^2.0.38"  # For database operations
python-dotenv = "^1.0.1"  # For environment management
langchain = "^0.3.25"  # For LLM chain operations
numpy = "^1.26.4"  # For numerical operations
pytest = "^8.4.2"  # For testing
openevals = "^0.1.0"
langgraph = "^0.2.70"
langchain-openai = "^0.3.17"
requests = "^2.32.3"


# 2025-7-6爬取了项目内所有的import和from,增加了如下依赖
pymilvus = "^2.5.12"  # Milvus client
llama-index = "*"  # LLM-based indexing
langsmith = "^0.4.27"  # Langchain logging tool
typing-extensions = "^4.12.2"  # For older Python compatibility
anthropic = "^0.51.0"  # For Anthropic models
langchain-anthropic = "^0.3.13"  # For Anthropic integration
langchain-community = "^0.3.24"  # For community extensions
langchain-core = "^0.3.76"  # Core langchain functionality
pandas = "^2.3.1"  # For data manipulation
tiktoken = "^0.8.0"  # For token counting

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"
sphinx = "^7.2.0"
sphinx-rtd-theme = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = ">=3.10,<3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
